refactor : 공통된 특징에 맞춰 패키지 분리
