+++ src/test/java/domain/FloorTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import common.exception.message.ExceptionMessage;
 import domain.bridge.strategy.BridgeGeneratorStub;
+import domain.ladder.Floor;
+import domain.ladder.LadderBridge;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
