+++ src/test/java/domain/PlayerNameTest.java
@@ -1,6 +1,7 @@
 package domain;
 
 import common.exception.message.ExceptionMessage;
+import domain.player.PlayerName;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
