+++ src/test/java/domain/PlayerNamesTest.java
@@ -1,6 +1,8 @@
 package domain;
 
 import common.exception.message.ExceptionMessage;
+import domain.player.PlayerName;
+import domain.player.PlayerNames;
 import java.util.List;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
