+++ src/test/java/domain/bridge/strategy/BridgeGeneratorStub.java
@@ -1,7 +1,7 @@
 package domain.bridge.strategy;
 
-import domain.LadderBridge;
-import domain.bridge.BridgeGenerator;
+import domain.ladder.LadderBridge;
+import domain.ladder.strategy.BridgeGenerator;
 import java.util.List;
 
 public class BridgeGeneratorStub implements BridgeGenerator {
